senate_stocks/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dj-database-url==0.5.0
Django==3.1.5
django-extensions==3.1.1
django-filter==2.4.0
django-heroku==0.3.1
django-localflavor==3.1
djangorestframework==3.12.2
future==0.18.2
greenlet==1.0.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==1.0.2
numpy==1.19.5
pandas==1.2.0
psycopg2==2.8.6
pynvim==0.4.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-stdnum==1.14
pytz==2020.5
requests==2.25.1
retrying==1.3.3
six==1.11.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
